ci: add uv tooling, ruff, and GitHub Actions
The CPython protocol library had tests but no shared
commit/lint tooling or CI, so style and type drift versus
the rest of the stack was invisible. Commitizen, pre-commit,
ruff, and a pytest workflow now match the other Python
repos. Typing is modernized and sources are formatted so
the new hooks pass.

README test instructions now use `uv run pytest`.

+++ src/simple485_remastered/utils.py
@@ -45,7 +45,7 @@ def get_milliseconds() -> int:
     Returns:
         int: The current time in milliseconds since the Epoch.
     """
-    return int(round(time.time() * 1000))
+    return round(time.time() * 1000)
 
 
 def microseconds_to_seconds(microseconds: int | float) -> float:
